refactor(skill-audit-ui): isolate audited mutation lifecycle
